+++ src/third_party_libs/protobuf/src/google/protobuf/compiler/cpp/cpp_generator.cc
@@ -80,7 +80,7 @@ bool CppGenerator::Generate(const FileDescriptor* file,
   // __declspec(dllimport) depending on what is being compiled.
   string dllexport_decl;
 
-  for (int i = 0; i < options.size(); i++) {
+  for (size_t i = 0; i < options.size(); i++) {
     if (options[i].first == "dllexport_decl") {
       dllexport_decl = options[i].second;
     } else {
